# Speed up template lookups by avoiding splats and ===