--- a/actionview/lib/action_view/path_set.rb
+++ b/actionview/lib/action_view/path_set.rb
@@ -44,28 +44,31 @@ module ActionView #:nodoc:
       METHOD
     end
 
-    def find(*args)
-      find_all(*args).first || raise(MissingTemplate.new(self, *args))
+    def find(path, prefixes, partial, details, details_key, locals)
+      find_all(path, prefixes, partial, details, details_key, locals).first ||
+        raise(MissingTemplate.new(self, path, prefixes, partial, details, details_key, locals))
     end
 
-    def find_all(path, prefixes = [], *args)
-      _find_all path, prefixes, args
+    def find_all(path, prefixes, partial, details, details_key, locals)
+      search_combinations(prefixes) do |resolver, prefix|
+        templates = resolver.find_all(path, prefix, partial, details, details_key, locals)
+        return templates unless templates.empty?
+      end
+      []
     end
 
-    def exists?(path, prefixes, *args)
-      find_all(path, prefixes, *args).any?
+    def exists?(path, prefixes, partial, details, details_key, locals)
+      find_all(path, prefixes, partial, details, details_key, locals).any?
     end
 
     private
-      def _find_all(path, prefixes, args)
+      def search_combinations(prefixes)
         prefixes = Array(prefixes)
         prefixes.each do |prefix|
           paths.each do |resolver|
-            templates = resolver.find_all(path, prefix, *args)
-            return templates unless templates.empty?
+            yield resolver, prefix
           end
         end
-        []
       end
 
       def typecast(paths)