Method calls involving splats are slower, and this previously did several levels of splats, which was also hard to grok (I'd struggled with this in the past). Though this PR adds a little repetition I think it's a net benefit (I also expect the definitions of `find`/`find_all`/`exists` to become more different in the future).

This also replaces a `===` check with a call to `Array()` which is faster and IMO cleaner.

<details>
<summary><b>Benchmark</b></summary>

``` ruby
require "benchmark/ips"
require "action_view"
require "action_pack"
require "action_controller"

class TestController < ActionController::Base
end

ActionView::Base.logger = nil

TestController.view_paths = [File.expand_path("actionview/test/fixtures")]
controller_view = TestController.new.view_context
LOOKUP_CONTEXT = TestController.new.lookup_context

result = Benchmark.ips do |x|
  x.config(:time => 5, :warmup => 2)

  x.report("find_all found") do
    # Single candidate
    LOOKUP_CONTEXT.find_all("basic", ["test"], false)
  end

  x.report("find_all missing") do
    LOOKUP_CONTEXT.find_all("no_template_here", ["test"], true)
  end

  x.report("exists missing") do
    LOOKUP_CONTEXT.exists?("no_template_here", ["test"], true)
  end

  x.report("find_all many results") do
    LOOKUP_CONTEXT.find_all("hello_world", ["test"], false)
  end
end
```

</details>

**Before**

```
$ be ruby lookup_benchmark.rb
Warming up --------------------------------------
      find_all found    44.100k i/100ms
    find_all missing    45.622k i/100ms
      exists missing    41.121k i/100ms
find_all many results
                        43.849k i/100ms
Calculating -------------------------------------
      find_all found    437.937k (± 0.4%) i/s -      2.205M in   5.035045s
    find_all missing    455.570k (± 0.3%) i/s -      2.281M in   5.007187s
      exists missing    410.674k (± 0.3%) i/s -      2.056M in   5.006581s
find_all many results
                        438.060k (± 0.3%) i/s -      2.192M in   5.004969s
```

**After**

```
$ be ruby lookup_benchmark.rb
Warming up --------------------------------------
      find_all found    47.950k i/100ms
    find_all missing    50.592k i/100ms
      exists missing    50.084k i/100ms
find_all many results
                        48.004k i/100ms
Calculating -------------------------------------
      find_all found    484.917k (± 0.3%) i/s -      2.445M in   5.043091s
    find_all missing    507.769k (± 0.5%) i/s -      2.580M in   5.081569s
      exists missing    502.863k (± 0.6%) i/s -      2.554M in   5.079695s
find_all many results
                        483.612k (± 0.6%) i/s -      2.448M in   5.062523s
```